+++ udata/tests/apiv2/test_datasets.py
@@ -181,3 +181,116 @@ class DatasetResourceAPIV2Test(APITestCase):
         assert data['page_size'] == DEFAULT_PAGE_SIZE
         assert data['next_page'] is None
         assert data['previous_page'] is None
+
+
+class DatasetExtrasAPITest(APITestCase):
+    modules = None
+
+    def setUp(self):
+        self.login()
+        self.dataset = DatasetFactory(owner=self.user)
+
+    def test_get_dataset_extras(self):
+        self.dataset.extras = {'test::extra': 'test-value'}
+        self.dataset.save()
+        response = self.get(url_for('apiv2.dataset_extras', dataset=self.dataset))
+        self.assert200(response)
+        data = response.json
+        assert data['test::extra'] == 'test-value'
+
+    def test_update_dataset_extras(self):
+        self.dataset.extras = {'test::extra': 'test-value', 'test::extra-second': 'test-value-second'}
+        self.dataset.save()
+
+        data = ['test::extra-second', 'another::key']
+        response = self.put(url_for('apiv2.dataset_extras', dataset=self.dataset), data)
+        self.assert400(response)
+        assert response.json['message'] == 'Wrong payload format, dict expected'
+
+        data = {'test::extra-second': 'test-value-changed', 'another::key': 'another-value'}
+        response = self.put(url_for('apiv2.dataset_extras', dataset=self.dataset), data)
+        self.assert200(response)
+
+        self.dataset.reload()
+        assert self.dataset.extras['test::extra'] == 'test-value'
+        assert self.dataset.extras['test::extra-second'] == 'test-value-changed'
+        assert self.dataset.extras['another::key'] == 'another-value'
+
+    def test_delete_dataset_extras(self):
+        self.dataset.extras = {'test::extra': 'test-value', 'another::key': 'another-value'}
+        self.dataset.save()
+
+        data = {'another::key': 'another-value'}
+        response = self.delete(url_for('apiv2.dataset_extras', dataset=self.dataset), data)
+        self.assert400(response)
+        assert response.json['message'] == 'Wrong payload format, list expected'
+
+        data = ['another::key']
+        response = self.delete(url_for('apiv2.dataset_extras', dataset=self.dataset), data)
+        self.assert204(response)
+
+        self.dataset.reload()
+        assert len(self.dataset.extras) == 1
+        assert self.dataset.extras['test::extra'] == 'test-value'
+
+
+class DatasetResourceExtrasAPITest(APITestCase):
+    modules = None
+
+    def setUp(self):
+        self.login()
+        self.dataset = DatasetFactory(owner=self.user)
+
+    def test_get_ressource_extras(self):
+        '''It should fetch a resource from the API'''
+        resource = ResourceFactory()
+        resource.extras = {'test::extra': 'test-value'}
+        self.dataset.resources.append(resource)
+        self.dataset.save()
+        response = self.get(url_for('apiv2.resource_extras', dataset=self.dataset,
+                                    rid=resource.id))
+        self.assert200(response)
+        data = response.json
+        assert data['test::extra'] == 'test-value'
+
+    def test_update_resource_extras(self):
+        resource = ResourceFactory()
+        resource.extras = {'test::extra': 'test-value', 'test::extra-second': 'test-value-second'}
+        self.dataset.resources.append(resource)
+        self.dataset.save()
+
+        data = ['test::extra-second', 'another::key']
+        response = self.put(url_for('apiv2.resource_extras', dataset=self.dataset,
+                                    rid=resource.id), data)
+        self.assert400(response)
+        assert response.json['message'] == 'Wrong payload format, dict expected'
+
+        data = {'test::extra-second': 'test-value-changed', 'another::key': 'another-value'}
+        response = self.put(url_for('apiv2.resource_extras', dataset=self.dataset,
+                                    rid=resource.id), data)
+        self.assert200(response)
+
+        self.dataset.reload()
+        assert self.dataset.resources[0].extras['test::extra'] == 'test-value'
+        assert self.dataset.resources[0].extras['test::extra-second'] == 'test-value-changed'
+        assert self.dataset.resources[0].extras['another::key'] == 'another-value'
+
+    def test_delete_resource_extras(self):
+        resource = ResourceFactory()
+        resource.extras = {'test::extra': 'test-value', 'another::key': 'another-value'}
+        self.dataset.resources.append(resource)
+        self.dataset.save()
+
+        data = {'another::key': 'another-value'}
+        response = self.delete(url_for('apiv2.resource_extras', dataset=self.dataset,
+                                       rid=resource.id), data)
+        self.assert400(response)
+        assert response.json['message'] == 'Wrong payload format, list expected'
+
+        data = ['another::key']
+        response = self.delete(url_for('apiv2.resource_extras', dataset=self.dataset,
+                                       rid=resource.id), data)
+        self.assert204(response)
+        self.dataset.reload()
+        assert len(self.dataset.resources[0].extras) == 1
+        assert self.dataset.resources[0].extras['test::extra'] == 'test-value'
